power-libperfmgr: Add a timer mechanism to load signal boost
Currently, load hint signals boost the first frame but rely on frames
reporting an actual duration to end that boost. This change adds a
secondary timer to provide a maximum boost time based on the target,
in case the session fails to report an actual duration or gets rate
limited.

This patch also refactors the existing timer into a generic superclass
to make sharing timer logic easier.

Test: manual